backendv2/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
httpx==0.25.1
clerk-backend-api==0.0.5
python-jose==3.3.0
pydicom==2.4.3
numpy==1.26.0
scipy==1.11.3
pydantic==2.5.0
python-dotenv==1.0.0
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.76
pillow==10.0.0 